general : debugging

+++ src/main/scala/com/muhalfian/spark/util/AggTools.scala
@@ -109,23 +109,9 @@ object AggTools {
     val aggregateArray = mongoRDD.map(r => {
       var tempJava = r.get("text_aggregate", new java.util.ArrayList[String]())
 
-      // var tempSeq = tempJava.map(row => {
-      //   (row.get(_1, new java.lang.Integer), row.get(_2, new java.lang.Double))
-      // }).toSeq
-      // println(tempSeq)
-      //   var word = row.drop(1).dropRight(1).split("\\,")
-      //   // var index = masterWordsIndex.indexWhere(_ == word(0))
-      //   // if(index == -1){
-      //   //   masterWordsIndex += word(0)
-      //   //   index = masterWordsIndex.size - 1
-      //   // }
-      //
-      //   (word(0).toInt, word(1).toDouble)
-      // }).toSeq
-
       var tempSeq = tempJava.map( row => {
         var word = row.drop(1).dropRight(1).split("\\,")
-        (word(0), word(1))
+        (word(0).toInt, word(1).toDouble)
       }).toSeq
 
       val vectorData = Vectors.sparse(size, tempSeq.sortWith(_._1 < _._1)).toDense.toArray
